AD binaries bootsrapping
AlfaDirectAPI is a shared library which
on Win
 loads ADAPI.dll
on Unix
 starts ADAPIServer.exe under wine which
   loads ADAPI.dll

the basic idea is to include all this binaries directly
into AlfaDirectAPI shared library, i.e. do compile-in.

this helps us to avoid many problems of searching directories
while AlfaDirectAPI loading, now all requirements are
compiled-in and AlfaDirectAPI does bootstraping on first
ADConnection::connect call.

PS:
 actually this patch is not tested on Windows.
 seems to me ADLocalLibrary can be broken.